refactor spec_helper
